+++ integration/integration_suite_test.go
@@ -6,15 +6,13 @@ import (
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-
-	"github.com/cloud-native-application/rudrx/pkg/server"
 )
 
 var restServer http.Handler
 
-var _ = BeforeSuite(func() {
-	restServer = server.SetupRoute()
-})
+//var _ = BeforeSuite(func() {
+//	restServer = server.SetupRoute()
+//})
 
 func TestIntegration(t *testing.T) {
 	RegisterFailHandler(Fail)
