+++ pkg/oam/env.go
@@ -1,0 +1,27 @@
+package oam
+
+import (
+	"encoding/json"
+	"fmt"
+	"io/ioutil"
+	"os"
+	"path/filepath"
+
+	"github.com/cloud-native-application/rudrx/api/types"
+	"github.com/cloud-native-application/rudrx/pkg/utils/system"
+)
+
+func GetEnvByName(name string) (*types.EnvMeta, error) {
+	data, err := ioutil.ReadFile(filepath.Join(system.GetEnvDirByName(name), system.EnvConfigName))
+	if err != nil {
+		if os.IsNotExist(err) {
+			return nil, fmt.Errorf("%s not exist", name)
+		}
+		return nil, err
+	}
+	var meta types.EnvMeta
+	if err = json.Unmarshal(data, &meta); err != nil {
+		return nil, err
+	}
+	return &meta, nil
+}
