+++ pkg/server/handler/appHandlers.go
@@ -1,6 +1,13 @@
 package handler
 
 import (
+	"net/http"
+
+	"github.com/cloud-native-application/rudrx/pkg/server/util"
+
+	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	"github.com/cloud-native-application/rudrx/pkg/oam"
 	"github.com/gin-gonic/gin"
 	ctrl "sigs.k8s.io/controller-runtime"
 )
@@ -26,6 +33,21 @@ func GetApps(c *gin.Context) {
 }
 
 func ListApps(c *gin.Context) {
+	kubeClient := c.MustGet("KubeClient")
+	envName := c.Param("envName")
+	envMeta, err := oam.GetEnvByName(envName)
+	if err != nil {
+		util.HandleError(c, util.StatusInternalServerError, err)
+	}
+	namespace := envMeta.Namespace
+
+	ctx := util.GetContext(c)
+	applicationMetaList, err := oam.RetrieveApplicationsByName(ctx, kubeClient.(client.Client), "", namespace)
+	if err != nil {
+		util.HandleError(c, util.StatusInternalServerError, err.Error())
+		return
+	}
+	c.JSON(http.StatusOK, applicationMetaList)
 }
 
 func DeleteApps(c *gin.Context) {
